+++ bindings/core/src/method_handler/wallet.rs
@@ -127,14 +127,11 @@ pub(crate) async fn call_wallet_method_internal(wallet: &Wallet, method: WalletM
         }
         WalletMethod::GenerateAddress {
             account_index,
-            internal,
             address_index,
             options,
             bech32_hrp,
         } => {
-            let address = wallet
-                .generate_address(account_index, internal, address_index, options)
-                .await?;
+            let address = wallet.generate_address(account_index, address_index, options).await?;
 
             let bech32_hrp = match bech32_hrp {
                 Some(bech32_hrp) => bech32_hrp,